requirements.txt
python-telegram-bot==20.7
yfinance==0.2.40
pandas==2.2.3
numpy==2.1.3
requests==2.31.0
